Added TimerOverlay Changed StopWatch and RestartTimer to extend TimerOverlay
